# server/swagger.yaml
...
openapi: 3.0.0
info:
  title: HoloTracker
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /employees/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Employees
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Employees
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employees"
    post:
      summary: CRUD ACTION update
      tags:
        - Employees
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employees"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employees"
  /employees/findBy_Farms/{key}:
    get:
      summary: CRUD ACTION findBy_Farms
      tags:
        - Employees
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employees"
  /employees/findBy_Plants/{key}:
    get:
      summary: CRUD ACTION findBy_Plants
      tags:
        - Employees
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employees"
  /employees/findBy_Product/{key}:
    get:
      summary: CRUD ACTION findBy_Product
      tags:
        - Employees
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employees"
    
  #-----  Farm  -----
...
  /farm/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Farm
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Farm
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
    post:
      summary: CRUD ACTION update
      tags:
        - Farm
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Farm"
  /farm/findBy_Distributer/{key}:
    get:
      summary: CRUD ACTION findBy_Distributer
      tags:
        - Farm
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
  /farm/findBy_Plants/{key}:
    get:
      summary: CRUD ACTION findBy_Plants
      tags:
        - Farm
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
    
  #-----  Plants  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Employees:
      properties:
        _id:
          type: string
        Address:
          type: string
        FirstName:
          type: string
        LastName:
          type: string
        Title:
          type: string
      required:
        -  _id
        - LastName
        - Title
    Farm:
      properties:
        _id:
          type: string
        Address:
          type: string
        Name:
          type: string
        Status:
          type: string
      required:
        -  _id
        - Name
...